micro: deepen workspace reducer seam
